Post funktio siirretty omaan tiedostoonsa
